Update gradle and create 4 UseCases

+++ app/src/main/java/ru/nsu/bashev/modules/database/account/IAccountDBHandler.java
@@ -8,12 +8,12 @@ import ru.nsu.bashev.modules.database.categories.CategoriesDBHandler;
 public interface IAccountDBHandler {
     CategoriesDBHandler getCategoriesDBHandler();
     void addAccount(Account account);
-    void updateAccount(int id, Account account);
-    Account getAccount(int id);
+    void updateAccount(long id, Account account);
+    Account getAccount(long id);
     List<Account> getAccountsByTitle(String titile);
     List<Account> getAccountsByEmail(Email email);
     List<Account> getAccountsByLogin(Login login);
     List<Account> getAllAccounts();
-    void deleteAccount(int id);
+    void deleteAccount(long id);
     void deleteAll();
 }
